Natural subsetting for indicies

Natural subsetting for indicies

Algebra within the raml ecosystem behaves exactly as you'd expect it to.

Totally redundent, but suppresses warnings when two variables are compared.

Totally redundent, but suppresses warnings when two variables are compared.

Totally redundent, but suppresses warnings when two variables are compared.

# S4 method for ramlArray
[(x, i, j, ..., drop = TRUE)

# S4 method for ramlVariable,numeric
+(e1, e2)

# S4 method for numeric,ramlVariable
+(e1, e2)

# S4 method for ramlVariable,ramlVariable
+(e1, e2)

# S4 method for ramlVariable,AffineExpr
+(e1, e2)

# S4 method for AffineExpr,numeric
+(e1, e2)

# S4 method for numeric,AffineExpr
+(e1, e2)

# S4 method for AffineExpr,ramlVariable
+(e1, e2)

# S4 method for AffineExpr,AffineExpr
+(e1, e2)

# S4 method for ramlVariable,numeric
*(e1, e2)

# S4 method for AffineExpr,numeric
*(e1, e2)

# S4 method for numeric,AbstractRamlAlgObject
*(e1, e2)

# S4 method for AbstractRamlAlgObject,numeric
-(e1, e2)

# S4 method for numeric,AbstractRamlAlgObject
-(e1, e2)

# S4 method for AbstractRamlAlgObject,AbstractRamlAlgObject
-(e1, e2)

# S4 method for AbstractRamlAlgObject,numeric
/(e1, e2)

# S4 method for ANY,AbstractRamlAlgObject
>=(e1, e2)

# S4 method for ANY,AbstractRamlAlgObject
<=(e1, e2)

# S4 method for ANY,AbstractRamlAlgObject
==(e1, e2)

# S4 method for AbstractRamlAlgObject,ANY
>=(e1, e2)

# S4 method for AbstractRamlAlgObject,ANY
<=(e1, e2)

# S4 method for AbstractRamlAlgObject,ANY
==(e1, e2)

# S4 method for AbstractRamlAlgObject,AbstractRamlAlgObject
==(e1, e2)

# S4 method for AbstractRamlAlgObject,AbstractRamlAlgObject
<=(e1, e2)

# S4 method for AbstractRamlAlgObject,AbstractRamlAlgObject
>=(e1, e2)

Arguments

x
The object of "ramlArray" to be subsetted.

i
The first index.

j
The second index (if needed).

...
Additional indicies.

drop
(Not used).

e1
The first algebraic object.

e2
The second algebraic object.

Examples

m <- Model()
m$var(x)
m$var(y)
x + x == 2 * x

#> Error in e2 * e1: invalid object (non-function) used as method

x - x == 0 * x

#> Found more than one class "ramlVariable" in cache; using the first, from namespace 'raml'

#> Also defined by ‘.GlobalEnv’

#> Found more than one class "ramlVariable" in cache; using the first, from namespace 'raml'

#> Also defined by ‘.GlobalEnv’

#> Error in e2 * e1: invalid object (non-function) used as method

x + y + x == x + x + y

#> 2*x + 1*y == 2*x + 1*y

m$objective(x + y)
m$constraint(x + 2 * y <= 3)

#> Error in e2 * e1: invalid object (non-function) used as method