A class that represents an affine combination of variables.

A class that represents an affine combination of variables.

Slots

lhs:

Object of class "AffineExpr", the left-hand side of the comparison.

rhs:

Object of class "AffineExpr", the rightt-hand side of the comparison.

comp:

Object of class "character", which indicates if the Comparison being made is ">=", "<=", or "==".